connect every class

+++ src/main/java/logic/item/potion/StrengthPotion.java
@@ -1,13 +1,15 @@
 package logic.item.potion;
 
 import logic.base.BasePotion;
+import logic.creatures.Player;
 
 public class StrengthPotion extends BasePotion {
-    public StrengthPotion(){
-        super("Strength Potion",1.5);
+    public StrengthPotion() {
+        super("Strength Potion", 1.5); // Increases Strength by 50%
     }
+
     @Override
-    public void consume(Player p){
-        p.setStrength(p.getStrength()*getStat());
+    public void consume(Player p) {
+        p.setStrength((int)(p.getStrength() * getStat()));
     }
 }
